rename node.py to worknode.py
