+++ aiida_workgraph/nodes/qe.py
@@ -1,5 +1,5 @@
 from typing import Dict
-from aiida_workgraph.node import WorkNode
+from aiida_workgraph.worknode import WorkNode
 
 
 class AiiDAKpoint(WorkNode):
